Implement first response test.
