+++ pom.xml
@@ -167,6 +167,13 @@
             <scope>test</scope>
         </dependency>
 
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>3.2.4</version>
+            <scope>test</scope>
+        </dependency>
+
         <!-- cucumber -->
         <dependency>
             <groupId>io.cucumber</groupId>
@@ -215,9 +222,11 @@
                   <includes>
                     <include>**/*Test*.*</include>
                   </includes>
+                  <!--
                   <excludes>
                     <exclude>**/*Cucumber*.*</exclude>
                   </excludes>
+                  -->
                 </configuration>
               </execution>
               <execution>
